requirements.txt
sentence-transformers
numpy
plotly
scikit-learn
umap-learn
langchain_text_splitters
pypdf
groq
